+++ class/actions_subtotal.class.php
@@ -203,7 +203,13 @@ class ActionsSubtotal
 
 					if (!empty($conf->global->SUBTOTAL_AUTO_ADD_SUBTOTAL_ON_ADDING_NEW_TITLE) && $qty < 10) TSubtotal::addSubtotalMissing($object, $qty);
 
-	    			TSubtotal::addSubTotalLine($object, $title, $qty);
+	    			if ($conf->global->MAIN_VIEW_LINE_NUMBER == 1) {
+						$rang = GETPOST('rank', 'int') ? (int) GETPOST('rank', 'int') : '-1';
+						$newlineid = TSubtotal::addSubTotalLine($object, $title, $qty, $rang);
+						echo '<div id="newlineid">'.$newlineid.'</div>';
+					} else {
+						TSubtotal::addSubTotalLine($object, $title, $qty);
+					}
 				}
 				else if($action==='ask_deleteallline') {
 						$form=new Form($db);
@@ -243,10 +249,13 @@ class ActionsSubtotal
 
 		$jsData = array(
 			'conf' => array(
-				'SUBTOTAL_USE_NEW_FORMAT' => !empty($conf->global->SUBTOTAL_USE_NEW_FORMAT)
+				'SUBTOTAL_USE_NEW_FORMAT' => !empty($conf->global->SUBTOTAL_USE_NEW_FORMAT),
+				'MAIN_VIEW_LINE_NUMBER' => !empty($conf->global->MAIN_VIEW_LINE_NUMBER),
+				'token' => ((float) DOL_VERSION < 11.0) ?  $_SESSION['newtoken'] : newToken()
 			),
 			'langs' => array(
-				'Level' => $langs->trans('Level')
+				'Level' => $langs->trans('Level'),
+				'Position' => $langs->transnoentities('Position')
 			)
 		);
 
@@ -273,7 +282,8 @@ class ActionsSubtotal
 					     $( "#dialog-prompt-subtotal" ).remove();
 
 						 var dialog_html = '<div id="dialog-prompt-subtotal" '+(action == 'addSubtotal' ? 'class="center"' : '')+' >';
-						 dialog_html += '<input id="token" name="token" type="hidden" value="<?php echo ((float) DOL_VERSION < 11.0) ?  $_SESSION['newtoken'] : newToken(); ?>" />';
+						 dialog_html += '<input id="token" name="token" type="hidden" value="' + jsSubTotalData.conf.token + '" />';
+
 
 						 if (typeof show_under_title != 'undefined' && show_under_title)
 						 {
@@ -293,12 +303,18 @@ class ActionsSubtotal
 							else dialog_html += '<input id="sub-total-title" size="30" value="" placeholder="'+label+'" />';
 						}
 
+						if (action == 'addSubtotal'){
+							dialog_html += '<input id="sub-total-title" size="30" value="" placeholder="'+label+'" />';
+						}
+
+						if(jsSubTotalData.conf.MAIN_VIEW_LINE_NUMBER) {
+							dialog_html += '&emsp;<input style="max-width: 80px;" id="subtotal_line_position" name="subtotal_line_position" type="number" min="0" step="1" size="1" text-align="right" placeholder="' + jsSubTotalData.langs.Position + '" />';
+						}
+
 						if (action == 'addTitle' || action == 'addSubtotal')
 						{
-							if (action == 'addSubtotal') dialog_html += '<input id="sub-total-title" size="30" value="" placeholder="'+label+'" />';
-
 							if(jsSubTotalData.conf.SUBTOTAL_USE_NEW_FORMAT){
-								dialog_html += '&nbsp;<select name="subtotal_line_level">';
+  								dialog_html += '&emsp;<select name="subtotal_line_level">';
 								for (var i=1;i<10;i++)
 								{
 									dialog_html += '<option value="'+i+'">'+ jsSubTotalData.langs.Level +' '+i+'</option>';
@@ -337,17 +353,33 @@ class ActionsSubtotal
 	                        buttons: {
 	                            "Ok": function() {
 	                            	if (typeof use_textarea != 'undefined' && use_textarea && typeof CKEDITOR == "object" && typeof CKEDITOR.instances != "undefined" ){ updateAllMessageForms(); }
+									params.rank = 0;
+									if($(this).find('#subtotal_line_position').length > 0){
+										params.rank = $(this).find('#subtotal_line_position').val();
+									}
+
 									params.title = (typeof CKEDITOR == "object" && typeof CKEDITOR.instances != "undefined" && "sub-total-title" in CKEDITOR.instances ? CKEDITOR.instances["sub-total-title"].getData() : $(this).find('#sub-total-title').val());
 									params.under_title = $(this).find('select[name=under_title]').val();
 									params.free_text = $(this).find('select[name=free_text]').val();
 									params.level = $(this).find('select[name=subtotal_line_level]').val();
 									params.token = $(this).find('input[name=token]').val();
 
+									let microtime = new Date();
+									url_to+="&microtime="+ microtime.getTime(); // to avoid # ancor blocking refresh by adding same rank as curent
+
 									$.ajax({
 										url: url_ajax
 										,type: 'POST'
 										,data: params
-									}).done(function() {
+										,dataType: "html"
+									}).done(function(response) {
+										if(jsSubTotalData.conf.MAIN_VIEW_LINE_NUMBER == 1) {
+											newlineid = $($.parseHTML(response)).find("#newlineid").text();
+											url_to = url_to + "&gotoline=" + params.rank + "#row-" + newlineid;
+										}
+										else {
+											url_to = url_to + "&gotoline=" + params.rank + "#tableaddline";
+										}
 										document.location.href=url_to;
 									});